chore: divide queryFn from component
