+++ src/components/AddTabModal/YoutubeSearchWrapper.tsx
@@ -1,47 +1,19 @@
 import { useQuery } from "@tanstack/react-query";
 import React, { useState } from "react";
 
+import { getYoutubeSearch } from "@/api/youtube";
 import useDebounce from "@/hooks/useDebounce/useDebounce";
-import { youtubeApi } from "@/lib/axios";
 
 import YoutubeSearchResult from "./YoutubeSearchResult";
 import { Input } from "../ui/input";
-
-interface YoutubeSearchSnippet {
-  title: string;
-  thumbnails: {
-    medium: {
-      url: string;
-    };
-  };
-}
-
-export interface YoutubeSearchItem {
-  id: {
-    videoId: string;
-  };
-  snippet: YoutubeSearchSnippet;
-}
 
 export default function YoutubeSearchWrapper() {
   const [keyword, setKeyword] = useState("");
   const debouncedKeyword = useDebounce(keyword, 1000);
 
-  const getYoutubeSearch = useQuery({
+  const youtubeSearchQuery = useQuery({
     queryKey: ["youtubeSearch", debouncedKeyword],
-    queryFn: async ({ queryKey }) => {
-      const [_, keyword] = queryKey;
-      const res = await youtubeApi.get("", {
-        params: {
-          key: process.env.NEXT_PUBLIC_YOUTUBE_KEY,
-          part: "snippet",
-          q: `${keyword} audio`,
-          maxResults: 5,
-          type: "video",
-        },
-      });
-      return res.data.items;
-    },
+    queryFn: getYoutubeSearch,
     enabled: debouncedKeyword.length > 0,
   });
 
@@ -54,7 +26,7 @@ export default function YoutubeSearchWrapper() {
   return (
     <>
       <Input placeholder="곡 제목을 입력해주세요" onChange={handleChange} value={keyword} />
-      <YoutubeSearchResult queryResult={getYoutubeSearch} />
+      <YoutubeSearchResult queryResult={youtubeSearchQuery} />
     </>
   );
 }
